backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy==1.24.3
requests==2.31.0
groq==0.4.2
langchain-groq==0.0.1
langchain-core==0.1.0
langgraph==0.0.20
python-dotenv==1.0.0
twilio==8.10.0
httpx==0.25.2 